requirement.txt
winrt-runtime==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Graphics.Capture==3.2.1
winrt-Windows.Graphics.Capture.Interop==3.2.1
winrt-Windows.Graphics.DirectX==3.2.1
winrt-Windows.Graphics.DirectX.Direct3D11.Interop==3.2.1
winrt-Windows.Graphics.Imaging==3.2.1
winrt-Windows.Storage==3.2.1
winrt-Windows.UI.Composition==3.2.1
winrt-Windows.UI.WindowManagement==3.2.1

paddleocr==3.2.0
paddlepaddle-gpu==3.2.0
